@@ -3,16 +3,17 @@
     "license": "proprietary",
     "type": "project",
     "require": {
-        "php": ">=7.2.5",
+        "php": "8.2.*",
         "doctrine/annotations": "^2.0",
         "doctrine/doctrine-bundle": "^2.13",
         "doctrine/doctrine-migrations-bundle": "^3.3",
         "doctrine/orm": "^2.20",
+        "sensio/framework-extra-bundle": "^6.2",
         "symfony/asset": "5.4.*",
         "symfony/cache": "5.4.*",
         "symfony/console": "5.4.*",
         "symfony/dotenv": "5.4.*",
-        "symfony/flex": "^1.1",
+        "symfony/flex": "^2.2",
         "symfony/form": "5.4.*",
         "symfony/framework-bundle": "5.4.*",
         "symfony/mailer": "5.4.*",
@@ -27,7 +28,7 @@
         "symfony/yaml": "5.4.*"
     },
     "require-dev": {
-        "rector/rector": "*",
+        "rector/rector": "^2.0",
         "symfony/dotenv": "^5.4",
         "symfony/maker-bundle": "^1.0",
         "symfony/phpunit-bridge": "^5.4",
