@@ -6,36 +6,31 @@ use App\Entity\Task;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 
 class TaskController extends AbstractController
 {
-    private EntityManagerInterface $entityManager;
-    private TaskRepository $taskRepository;
-
     public function __construct(
-        EntityManagerInterface $entityManager,
-        TaskRepository $taskRepository
+        private EntityManagerInterface $entityManager,
+        private TaskRepository $taskRepository
     ) {
         $this->entityManager = $entityManager;
         $this->taskRepository = $taskRepository;
     }
 
 
-    /**
-     * @Route("/tasks", name="task_list")
-     */
-    public function list()
+    #[Route(path: '/tasks', name: 'task_list')]
+    public function list(): Response
     {
         return $this->render('task/list.html.twig', ['tasks' => $this->taskRepository->findAll()]);
     }
 
-    /**
-     * @Route("/tasks/create", name="task_create")
-     */
-    public function create(Request $request)
+    #[Route(path: '/tasks/create', name: 'task_create')]
+    public function create(Request $request): Response
     {
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
@@ -54,17 +49,15 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/tasks/{id}/edit", name="task_edit")
-     */
-    public function edit(Task $task, Request $request)
+    #[Route(path: '/tasks/{id}/edit', name: 'task_edit')]
+    public function edit(Task $task, Request $request): Response
     {
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->entityManager->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
@@ -77,10 +70,8 @@ class TaskController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/tasks/{id}/toggle", name="task_toggle")
-     */
-    public function toggleTask(Task $task)
+    #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle')]
+    public function toggleTask(Task $task): RedirectResponse
     {
         $task->toggle(!$task->isDone());
         $this->entityManager->flush();
@@ -90,10 +81,8 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list');
     }
 
-    /**
-     * @Route("/tasks/{id}/delete", name="task_delete")
-     */
-    public function deleteTask(Task $task)
+    #[Route(path: '/tasks/{id}/delete', name: 'task_delete')]
+    public function deleteTask(Task $task): RedirectResponse
     {
         $this->entityManager->remove($task);
         $this->entityManager->flush();
